lambda/requirements.txt
Pillow
pyzbar
opencv-python-headless
numpy
pydantic
boto3
